many fixes

+++ ledger-api/src/main/java/ee/taltech/ledger/api/dto/BlockDTO.java
@@ -8,6 +8,6 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BlockDTO {
-  private int transaction;
+  private double transaction;
   private String message;
 }
